backend/pyproject.toml
[tool.poetry]
name = "management-web-system-backend"
version = "0.1.0"
description = "Backend for Management Web System"
authors = ["nampq"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "src/**/*.py" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.37"
pydantic = {extras = ["email"], version = "^2.10.5"}
mysql-connector-python = "^9.1.0"
pydantic-settings = "^2.7.1"
sqlmodel = "^0.0.22"
itsdangerous = "^2.2.0"
pyjwt = "^2.10.1"
alembic = "^1.14.0"
asyncmy = "^0.2.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"
redis = "^5.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
